+++ src/stable/commands/deletecrowns.ts
@@ -85,7 +85,8 @@ class DeleteCrownCommand extends Command {
         time: 30000,
       }
     );
-    msg.delete();
+    const message_exists = message.channel.messages.cache.get(msg.id);
+    if (message_exists) msg.delete();
     if (reactions.size > 0) {
       let delete_stats;
       if (global) {
